Decouple Hostname Guessing from NetworkInit
